+++ nacc_attribute_deriver/attributes/missingness/modules/uds/__init__.py
@@ -11,10 +11,12 @@ from .missingness_b2 import *
 from .missingness_b3 import *
 from .missingness_b5 import *
 from .missingness_b6 import *
+from .missingness_b7 import *
 from .missingness_b8 import *
 from .missingness_b9 import *
 from .missingness_cx import *
 from .missingness_d1 import *
 from .missingness_d1a import *
 from .missingness_d1b import *
+from .missingness_d1c import *
 from .missingness_uds import *